+++ src/compiler/Lexer.cpp
@@ -38,7 +38,7 @@ Lexer::Lexer(istream &file) : infile(file), peekDone(false) {}
 #define _BC (buf[state.bufp])
 #define EOB (state.bufp >= buf.length())
 
-size_t Lexer::fillBuffer() {
+int64_t Lexer::fillBuffer() {
     if (!infile.good()) { // file stream
         return -1;
     }
